% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aquaman.R
\docType{data}
\name{Aquaman}
\alias{Aquaman}
\title{Box office from the movie Aquaman}
\format{
A data frame with 20 rows and variables \code{Date}, \code{Earnings} (in USD), and the count of the \code{Weekend}.
}
\source{
Beth Schaubroeck at USAFA.
Scraped from \url{https://www.boxofficemojo.com/release/rl3108800001/weekend/?ref_=bo_rl_tab}.
}
\usage{
Aquaman
}
\description{
Earnings each weekend after the release of the movie \emph{Aquaman}
}
\keyword{datasets}
