% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Body_fat.R
\docType{data}
\name{Body_fat}
\alias{Body_fat}
\title{Measurements of body-fat percentage and related quantities}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 252 rows and 14 columns.
}
\usage{
Body_fat
}
\description{
It's hard to measure body fat. This data frame records a direct
measure of body-fat percentage along with easier-to-make measurements.
The question of interest in collecting the data is whether there
is a way to estimate body-fat percentage from the easier-to-measure
quantities.
}
\keyword{datasets}
