% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boyle.R
\docType{data}
\name{Boyle}
\alias{Boyle}
\title{Robert Boyle's pressure vs volume measurements}
\format{
A data frame with 7 rows and numerical variables \code{pressure} and \code{volume}.
}
\source{
Boyle's notebooks are preserved at the Royal Society in London. The data in the \code{Boyle} data frame have been copied from \verb{https://chem.libretexts.org/Courses/University_of_California_Davis/UCD_Chem_002A/UCD_Chem_2A/Text/Unit_III\%3A_Physical_Properties_of_Gases/06.03_Relationships_among_Pressure\%2C_Temperature\%2C_Volume\%2C_and_Amount}
}
\usage{
Boyle
}
\description{
Pressure versus volume of air at a constant temperature as collected by Robert Boyle. Units are arbitrary.
}
\details{
Robert Boyle (1627-1691) was a founder of modern chemistry
and the scientific method in general. As any chemistry student already knows,
Boyle sought to understand the properties of gasses.
His results are summarized in \emph{Boyle's Law}.
The data frame \code{Boyle} contains two variables from one of Boyle's
experiments as reported in his lab notebook: pressure in a bag of
air and volume of the bag. The units of pressure are mmHg and the units of
volume are cubic inches.
}
\keyword{datasets}
