% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musical-notes.R
\docType{data}
\name{Cello}
\alias{Cello}
\alias{Cello_seg}
\alias{Violin}
\alias{Violin_seg}
\title{Short recordings of a cello and a violin}
\format{
Data frames (for convenience) with many rows and these variables:
\itemize{
\item \code{y} the amplitude of the sound wave (arbitrary units)
\item \code{t} the time of the sample in seconds.
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3616 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43776 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3638 rows and 2 columns.
}
\usage{
Cello

Cello_seg

Violin

Violin_seg
}
\description{
Only one note is involved.
\itemize{
\item \code{Cello}: a complete note
\item \code{Violin}: a complete note
\item \code{Cello_seg}: a brief snipet of sound after the initial transient
\item \code{Violin_seg}: as in \code{Cello_seg}
}
}
\keyword{datasets}
