% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chirps.R
\docType{data}
\name{Chirps}
\alias{Chirps}
\title{Cricket chirp rate and temperature}
\format{
A data frame with 55 rows
\itemize{
\item{\code{date}} {Which year the estimate is for.}
\item{\code{time}} {in 24-hour format}
\item{\code{chirps}} {the number of chirps per 15 seconds}
\item{\code{temperature}} {in degrees F}
}
}
\usage{
data(Chirps)
}
\description{
The rate at which crickets chirp is related to the ambient temperature. These data
are motivated and described at \url{https://www.globe.gov/explore-science/scientists-blog/archived-posts/sciblog/index.html_p=45.html}
}
\keyword{datasets}
