% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_plot.R
\name{contour_plot}
\alias{contour_plot}
\title{Contour plots of functions of two variables}
\usage{
contour_plot(
  object,
  formula,
  domain,
  npts = 100,
  labels = TRUE,
  filled = TRUE,
  contours_at = NULL,
  n_contours = 10,
  n_fill = 50,
  alpha = 1,
  fill_alpha = 0.1,
  label_alpha = 1,
  contour_color = NULL,
  label_color = ifelse(!is.null(contour_color), contour_color, "black"),
  skip = 1,
  guide = FALSE,
  guide_title = "output",
  color_scale = scale_color_viridis_c(),
  fill_scale = scale_fill_viridis_c(),
  ...
)
}
\arguments{
\item{object}{(Can be ignored. Used to pass previous layers into  this one.)}

\item{formula}{A tilde formula with two independent variables, e.g. \code{x*y~x+y}}

\item{domain}{A list giving the ranges for the two independent variables. See examples.}

\item{npts}{Integer number of points on each axis at which to evaluate the function}

\item{labels}{Logical flag: label the contours}

\item{filled}{Logical flag: fill between the contours}

\item{contours_at}{Vector of numbers. Contours will be drawn
at these levels of the output.}

\item{alpha}{Transparency of contours in \link{0-1}}

\item{fill_alpha}{Transparency of fill}

\item{label_alpha}{Likewise, for contour labels}

\item{contour_color}{Set to a string, e.g. \code{"blue"} for contours
to be drawn in a fixed color}

\item{label_color}{Defaults to \code{contour_color}. Can be set to a}

\item{skip}{Small integer. Skip this many contours between labeled contours}

\item{guide}{Logical flag, whether to show a color guide.}

\item{guide_title}{Character string  for title of guide (if any)}

\item{color_scale}{See ggplot2 color scales.}

\item{fill_scale}{See ggplot2 fill scales

to a fixed color.}

\item{nfill}{Integer number of points for calculating fill}
}
\description{
Creats a ggplot2-compatible contour plot of a function
of two variables.
}
