% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid_US.R
\docType{data}
\name{Covid_US}
\alias{Covid_US}
\title{COVID data from the first half of the pandemic}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 575 rows and 36 columns.
}
\usage{
Covid_US
}
\description{
For documentation, see \verb{https://covid19datahub.io/articles/doc/data.html}
}
\details{
See Guidotti, E., Ardia, D., (2020), "COVID-19 Data Hub", Journal
of Open Source Software 5(51):2376, doi: 10.21105/joss.02376.
}
\keyword{datasets}
