% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{domain}
\alias{domain}
\title{Specify a domain over which a function is to be graphed}
\usage{
domain(...)
}
\arguments{
\item{\dots}{One or more expression of the form \code{x = -5:5}}
}
\value{
a list with one component for each element in \dots
}
\description{
domain() is used with slice_plot(), contour_plot(), or interactive_plot()
to describe the plotting region. There is a standard syntax for
domain() (see the first example) but there are also shortcuts.
In the shorthand syntaxes, you can but don't have to specify the name of
the input. If it's not specified, the plotting programs will try
to do something sensible. But better to specify the names explicitly.
}
\details{
The colon operator is masked so that, for instance,
\code{x = 0.5:1.3} literally means "0.5 to 1.3", and not just 0.5 as the
base colon operator would give.
}
\examples{
\dontrun{
contour_plot(x / y ~ x + y, domain(x=-5:5, y=1:4))
slice_plot(x^2 ~ x, domain(x = 2.5:4.2)) # overrides colon operator
slice_plot(x^2 ~ x, domain(x = c(2.5, 4.2)))
slice_plot(x^2 ~ x, domain(x = 1 \%pm\% 0.5))
}

}
