% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doodle_fun.R
\name{doodle_fun}
\alias{doodle_fun}
\title{Construct a random function that is smooth}
\usage{
doodle_fun(vars = ~x & y, seed = NULL, n = 0)
}
\arguments{
\item{vars}{A right-sided tilde expression (like \code{~ x & y}) that indicates how many
inputs there should be and their names.}

\item{seed}{An integer setting the seed for random number generation.}

\item{n}{A positive integer. If non-zero, this is a hint about how many local
maxes and mins there should be. Just a hint.}
}
\description{
Sometimes it's nice to have a simple, smooth but curving function
to illustrate some mathematical point or to support an exercise. \code{doodle_fun()}
generates such functions. The function support tends to be in -10 to 10. For the
mathematically adept ... the functions are linear combinations of gaussians
with randomly selected centers, widths, and random coefficients. The \code{mosaic::rfun()} function
is very similar.
}
\examples{
# Two functions of two variables. They are different because the seeds are different.
f <- doodle_fun(~ x & y, seed = 101)
g <- doodle_fun(~ z & u, seed = 121)
}
