% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findZeros.R, R/fitModel.R
\name{findZeros}
\alias{findZeros}
\title{Find zeros of a function}
\description{
This is defined in the \verb{mosaic package}:
See \code{\link[mosaic]{findZeros}} for finding zeros of a function
of one variable
and \code{\link[mosaic]{findZerosMult}} for finding zeros of functions of two or more variables.

This is defined in the \verb{mosaic package}:
See \code{\link[mosaic]{findZeros}}
}
