% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFun.R
\name{graphFun}
\alias{graphFun}
\title{Graph a function of one or two variables}
\usage{
graphFun(formula, ..., labels = "top.pieces", fill = TRUE,
  alpha.tile = 0.4)
}
\description{
Graph a function of one or two variables
}
