% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunSurface.R
\name{graphFunSurface}
\alias{graphFunSurface}
\title{Graph function of two variables as a surface}
\usage{
graphFunSurface(formula, ...)
}
\arguments{
\item{formula}{of the sort accepted by `makeFun()`}

\item{...}{limits for the plot. See examples.}
}
\description{
Graph function of two variables as a surface
}
