% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Home_utilities.R
\docType{data}
\name{Home_utilities}
\alias{Home_utilities}
\title{Gas  and electricity usage by a home in St. Paul, MN}
\format{
A data.frame object with one row for each month
\itemize{
\item \code{month} - the month covered by the bill as a number
\item \code{day} - the terminating day of the bill
\item \code{year} - the year in which the month fell
\item \code{temp} -  average temperature over the entire  month
\item \code{kwh} - electricity usage in kilowatt-hours
\item \code{ccf} - natural gas usage in cubic feet
\item \code{thermsPerDay} - another measure of natural gas usage
\item \code{dur} - duration of the period covered by the bill, in days (typically about 30)
\item \code{totalbill} - dollar amount of the bill
\item \code{gasbill}: - dollar amount of the natural gas  part of the bill
\item \code{elecbill} - dollar amount of the electricity bill
\item \code{generatedKwh} - Energy generated by the photovoltaic system
\item \code{notes} - miscellaneous  comments on the month's bill or changes in the house
}
}
\source{
Personal communication from the homeowner
}
\usage{
data(Home_utilities)
}
\description{
Gas  and electricity usage by a home in St. Paul, MN
}
\details{
A college mathematics professor teaching in St. Paul, Minnesota, USA collected these records
of his utility  bills starting in 1999. Starting in June 2022,
solar photovoltaics were installed on the roof of the house. There are months missing, a few transcription errors, etc.
\strong{These data will be updated occasionally, as new bills come in.}
}
\keyword{datasets}
