% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inscribed_circle.R
\name{inscribed_circle}
\alias{inscribed_circle}
\title{Create a data frame for a circle marking the curvature of a function.}
\usage{
inscribed_circle(ftilde, x0 = 0)
}
\arguments{
\item{ftilde}{A tilde expression defining a function}

\item{x0}{The value for the input at which the curvature is to be calculated.}
}
\value{
\code{inscribed_circle()} return a data frame with \code{x} and \code{y} components describing the inscribed circle
located at the point $(x_0, f(x_0))$. Plot this with \code{gf_path(y ~ x)}
\code{curvature_function()} returns a function that gives the curvature of the specified
function for any input $x$. It's much like \code{D()} in the way it is used.
}
\description{
Create a data frame for a circle marking the curvature of a function.
}
