% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_in_domain.R
\name{is_in_domain}
\alias{is_in_domain}
\title{check whether a value is in a domain}
\usage{
is_in_domain(val, domain, ...)
}
\arguments{
\item{val}{the values to be checked}

\item{domain}{the domain}

\item{\dots}{additional domains}
}
\description{
A convenience function to see if a value is inside
(or on the boundaries of) a domain or a set of domain segments.
In the case of multiple segments, the check is
whether \code{val} is within any of them.
}
\examples{
is_in_domain(1:10, domain(x=5.5:8.5), domain(x=1:2))
is_in_domain(mtcars, domain(mpg=15:20, wt=2:4.5), domain(mpg=25:28))
}
