% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M2014F.R
\docType{data}
\name{M2014F}
\alias{M2014F}
\title{Mortality versus age for females in the US in 2014}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 120 rows and 5 columns.
}
\usage{
M2014F
}
\description{
These data are from the US Social Security Administration.
}
\details{
\itemize{
\item \code{pmort} the probability of a person of the given age dying in 2014
\item \code{nliving} traces a hypothetical population of 100,000 people through the mortality at each age.
\item \code{died} for the hypothetical population the number of people at that age who
survived to the age and then died at that age.
Source: \url{https://www.ssa.gov/oact/HistEst/PerLifeTables/2017/PerLifeTables2017.html}
}
}
\keyword{datasets}
