% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline.R
\name{simplify_fun}
\alias{simplify_fun}
\alias{inline_expr}
\alias{replace_arg_in_expr}
\title{Turn a 1-line function into an inline formula}
\usage{
simplify_fun(fun)

inline_expr(ex, old, new, env)

replace_arg_in_expr(ex, old, new)
}
\arguments{
\item{fun}{a function, such as produced by \code{makeFun()}.}

\item{ex}{an expression}

\item{old}{the name to be replaced as produced by as.name()}

\item{new}{the name to be substituted in for arg_name}

\item{env}{the environment for the expression}

\item{tilde}{a tilde expression as used by \code{makeFun()}.}
}
\description{
Turn a 1-line function into an inline formula
}
\examples{
infer_RHS(x^2 ~ x) # Formula is unchanged
}
