% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_energy.R
\name{PE_fun1}
\alias{PE_fun1}
\alias{PE_fun2}
\title{Potential energy functions used as examples in \emph{MOSAIC Calculus}.}
\usage{
PE_fun1(x1, y1)

PE_fun2(x1, y1, x2, y2, x3, y3)
}
\arguments{
\item{x1, y1, x2, y2, x3, y3}{coordinates of the masses.}
}
\description{
These appear in the "Optimization & constraint" chapter.
}
