% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{data}
\name{penguins}
\alias{penguins}
\title{Imports from other packages}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 344 rows and 8 columns.
}
\usage{
penguins
}
\description{
Imports from other packages
}
\keyword{datasets}
