% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Picket.R
\name{Picket}
\alias{Picket}
\title{Creates a "picket fence" of points for illustrating numerical integration}
\usage{
Picket(
  domain,
  h = 0.1,
  method = c("left", "right", "center", "trapezoid", "gauss")
)
}
\arguments{
\item{domain}{domain of integration as used in \code{slice_plot()}, \code{Integrate()}, and the similar mosaicCalc functions}

\item{h}{number giving the width between pickets. Could also have been called \code{dt} or \code{dx}, and so on.}

\item{method}{determines the weights for each element in the picket}
}
\description{
Creates a "picket fence" of points for illustrating numerical integration
}
