% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Planet_solar.R
\docType{data}
\name{Planet_solar}
\alias{Planet_solar}
\title{Solar irradiance of the planets}
\format{
A data frame with 9 rows
\itemize{
\item{\code{planet}} {Name of the planet (or planetoid)}
\item{\code{distance}} {distance from the sun in millions of km}
\item{\code{irradiance}} {solar brightness (in what units?)}
}
}
\usage{
data(Planet_solar)
}
\description{
Brightness of the sun at the various planets
}
\keyword{datasets}
