% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Robot_stations.R
\docType{data}
\name{Robot_stations}
\alias{Robot_stations}
\title{Waypoints on the path of a fictitious robot}
\format{
A data frame with 16 cases, each of which is an x-y location at the indicated time.
\itemize{
\item x,y X-Y location in mm
\item t time in seconds
}
}
\usage{
Robot_stations
}
\description{
This is used in \emph{MOSAIC Calculus} examples about splining.
}
\keyword{datasets}
