% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runners.R
\docType{data}
\name{Runners}
\alias{Runners}
\title{Running times}
\format{
Data frame with 24,334 rows
\itemize{
\item \code{year} the occasion on which the runner ran the Cherry Blossom Ten Miler
\item \code{age} the runners age at the time of that occasion
\item \code{gun} the time (in minutes) from the gun to crossing the finish line.
\item \code{net} time between crossing the start line and crossing the finish line. Sometimes this is missing. It's typically smaller than \code{gun} because it takes time to reach the starting line.
\item \code{sex} as recorded by the race organizers.
\item \code{previous} How many previous occasions did this person run the race.
\item \code{nruns} The total number of runs recorded for this person (including future runs)
\item \code{start_position} A qualitative indication of how close to the front of the pack the person was positioned at the time the gun went off.
}
}
\source{
Scraped from the Cherry Blossom web site by patient, unremunerated toil of the author.
}
\usage{
Runners
}
\description{
A longitudinal record of running times. There are 5977 individual runners included
here, each with a unique \code{id}. Each runner ran the 10-mile race (See \code{mosaicData::TenMileRace})
on multiple occasions.
}
\author{
Daniel Kaplan
}
\keyword{datasets}
