% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\docType{data}
\name{shapes}
\alias{shapes}
\title{Shapes in the book}
\format{
An object of class \code{list} of length 4.
}
\usage{
shapes
}
\description{
Used as examples for moment of inertial calculations
}
\keyword{datasets}
