% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleYacasIntegrate.R
\name{simpleYacasIntegrate}
\alias{simpleYacasIntegrate}
\title{Interface to integration using Ryacas}
\usage{
simpleYacasIntegrate(tilde, ...)
}
\arguments{
\item{tilde}{The tilde expression for the integration}

\item{\dots}{Bindings for parameters in \code{tilde}}
}
\description{
Interface to integration using Ryacas
}
