% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_plot.R
\name{slice_plot}
\alias{slice_plot}
\title{Plot a function of a single variable}
\usage{
slice_plot(
  object,
  formula,
  domain,
  npts = 100,
  label_text = "",
  label_x = 1,
  label_vjust = "top",
  ...
)
}
\arguments{
\item{object}{(Not intended for user. Receives the input from  any previous graphics layers.)}

\item{formula}{A tilde formula with one independent variable. All parameters must
be assigned specific numerical values.}

\item{domain}{A named list giving the extent of the x-axis scale. If there
is a previous plot layer using the same independent variable name, the
domain can be omitted and will be inferred from the previous layer.}

\item{npts}{Integer, number of points at which to evaluate the function.}

\item{label_text}{character string label to place near the graph curve. Default: none.}

\item{label_x}{number between 0 and 1 indicating the horizontal placement  of the \code{label_text}.

Additional arguments will be passed to \code{geom_line()}. Use, e.g. \code{color="red"}}
}
\description{
In a slice plot, there  is one independent variable. The graph
shows the output of the function versus the independent variable. It's
called a slice plot to distinguish it from a contour plot, in which
the graph has one axis for each independent variable and the output
of the function is shown by color and labels.
}
\examples{
\dontrun{
slice_plot(sin(x) ~ x, domain(x = range(-5, 15)))
f <- makeFun(sin(2*pi*t/P) ~ t)
slice_plot(f(t, P=20) ~ t, domain(t = range(-5, 10)), label_text = "Period 20", label_x=0.9)
}

}
