% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothers.R
\name{smoother}
\alias{smoother}
\alias{linearModel}
\alias{interpolating.function}
\alias{spliner}
\alias{connector}
\title{Functions from data}
\usage{
smoother(formula, data, span = 0.5, degree = 2, ...)

linearModel(formula, data, ...)

interpolating.function(formula, data, method = "fmm",
  monotonic = FALSE, connect = FALSE)

spliner(formula, data, method = "fmm", monotonic = FALSE)

connector(formula, data, method = "linear")
}
\arguments{
\item{formula}{A `y ~ x` type formula referring to the input and output variables in the data}

\item{data}{a data frame}

\item{span}{a hyperparameter for smoothers describing the amount of smoothing. (Default: 0.5)}

\item{degree}{a hyperparameter for smoothers,}

\item{...}{other arguments (not used?)}

\item{method}{character string describing the specific splining method}

\item{monotonic}{logical. Should the spline be monotonic?}
}
\description{
Functions that are derived from data
}
