% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSA_2007.R
\docType{data}
\name{SSA_2007}
\alias{SSA_2007}
\title{US Mortality table from 2007}
\format{
A data frame with 240 rows. Each row corresponds
to one age year from 0 to 120.
\itemize{
\item \code{Age}: The age year. 0 corresponds to birth through one year of age.
\item \code{Sex}: The sex for which the row is relevant.
\item \code{Mortality}: The fraction of people of that age who died in 2007.
\item \code{LifeExpectancy}: The calculated "life expectancy" at that age for that sex.
}
}
\usage{
SSA_2007
}
\description{
Mortality table from the US Social Security Adminstration
issued in 2007.
}
\details{
Life expectancy is a measure constructed from
a simulation. Start with 100,000 people at birth. Use the
mortality at each age to follow the living through successive
ages. The "life expectancy at birth" (age 0) is the average
age at death of those 100,000 people. For the life expectancy in
year n, consider only that fraction of the original 100,000 who survived to
year n. The life expectancy will be the average time until death for those survivors.
}
\keyword{datasets}
