% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbound.R
\name{unbound}
\alias{unbound}
\alias{bind_params}
\title{Identifying unbound inputs to a function}
\usage{
unbound(f)

bind_params(f, ...)
}
\arguments{
\item{f}{a function (not a tilde expr.)}

\item{\dots}{bindings for parameters}
}
\description{
\code{unbound()} finds if there are any unbound parameters in a function. This can be
useful for checking before handing a function over to a numerical routine.
\code{bind_params()} lets you add parameter bindings or override existing ones.
}
