% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/US_income.R
\docType{data}
\name{US_income}
\alias{US_income}
\title{Income distribution data from the US in 2009}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 2 columns.
}
\source{
La Haye and Zizler (2021) "The Lorenz Curve in the Classroom", \emph{The American Statistician}, 75(2):217-225]

Variables:
\itemize{
\item \code{income} The fraction of total household income that goes to the corresponding
poorest fraction of the population.
\item \code{pop} The fraction of the population whose aggregated income is reported
as the \code{income} value.
}
}
\usage{
US_income
}
\description{
Data such as these are used to study income inequality. One measure of
this is the "Gini coefficient," which can be estimated from a Lorenz
function fitted to such data.
}
\keyword{datasets}
