% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argM.R
\name{vector_arg}
\alias{vector_arg}
\title{convert a function with separate arguments
into one with a single vector argument
For use with optim.}
\usage{
vector_arg(f)
}
\arguments{
\item{f}{a function with multiple arguments}
}
\description{
convert a function with separate arguments
into one with a single vector argument
For use with optim.
}
