% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{vec}
\alias{vec}
\alias{rvec}
\alias{M_cols}
\alias{data_M}
\alias{veclen}
\alias{\%cang\%}
\alias{Q}
\alias{R_for_Q}
\alias{vbind}
\alias{vec_names}
\alias{\%dot\%}
\alias{\%onto\%}
\alias{\%perp\%}
\alias{normalize}
\alias{as_magnitude}
\alias{rmat}
\title{Utilities for vector/matrix calculations}
\usage{
vec(...)

rvec(n = 3L, fun = rnorm, ...)

M_cols(M, fun = "center")

data_M(.data, tilde)

veclen(v)

v \%cang\% w

Q(A)

R_for_Q(A)

vbind(..., deparse.level = 1)

vec_names(x, value)

u \%dot\% b

b \%onto\% A

b \%perp\% A

normalize(A)

as_magnitude(A, metric = c("2", "O", "I", "F", "M"))

rmat(nrow = 3, ncol = 6, fun = rnorm)
}
\arguments{
\item{fun}{either a function producing random numbers or, for \code{M_cols()}
the name of action to perform on each column.}

\item{A}{a matrix}

\item{u}{a row vector, but a column vector is acceptable too}

\item{b}{a column vector}

\item{metric}{metric to use for matrix norm}

\item{nrow}{number of rows for matrix}

\item{ncol}{number of columns for matrix}

\item{\ldots}{values to put in \code{vec()} \strong{or} params for random
number generator}
}
\description{
Convenience functions for basic operations relating to vector projection. These use
the \emph{MOSAIC Calc} conventions that require vectors to be one column matrices.
}
\details{
\itemize{
\item \code{vec(1,2,3)} makes a \strong{column} vector
\item \code{rvec(n)} random vector with \if{html}{\out{<n>}} elements
\item \code{vec_names(v, c("one", "two", ...))} gives names to the
vectors in a matrix.
\item \code{vbind()} collects column vectors into a matrix. Will also concatenate
matrices by column.
\item \code{M_cols()} operates on the columns of a matrix one-by-one. Operations
include \code{"center"}, \code{"unit"}, \code{"unitcenter"}, which scale the columns, and
\code{"mean"}, \code{"var"}, \code{"vlen"} which return a scalar for each column.)
\item \verb{\%dot\%} infix operator for dot product. Raw R vectors will work
\item \verb{\%cang\%} like \verb{\%dot\%} but returns the cosine of the angle
\item \verb{\%onto} infix projects left arg onto space (vector or matrix) defined
by right argument.
\item \verb{\%perp\%} infix. Like \verb{\%onto\%} but returns the reciprocal.
\item \code{vbind()} collects column vectors into a matrix
\item \code{Q(A)} the Q part of a qr decomposition
\item \code{R_from_Q(A)} the R part of a qr decomposition. Can apply
directly to the output of \code{Q(A)}
}

Convenience functions for basic operations relating to vector projection. These use
the \emph{MOSAIC Calc} conventions that require vectors to be one column matrices.
}
