% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vowels.R
\docType{data}
\name{Vowel_ee}
\alias{Vowel_ee}
\alias{Vowel_oh}
\alias{Oh_sound}
\alias{Ee_sound}
\title{Recordings of vowel sounds}
\format{
Data frames with two columns
\itemize{
\item \code{y}: The instantaneous amplitude of the vowel sound.
\item \code{t}: Time (seconds)
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 780 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7192 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10194 rows and 2 columns.
}
\usage{
Vowel_ee

Vowel_oh

Oh_sound

Ee_sound
}
\description{
"Oh" refers to "o" as in "stone." "Ee" to "eel".
}
\details{
\itemize{
\item \code{Oh_sound} and \code{Ee_sound} contain the complete utterances.
\item \code{Vowel_ee} and \code{Vowel_oh} are snippets of about 0.04 seconds duration.
}
}
\keyword{datasets}
