% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boyle.R, R/World_oil.R
\docType{data}
\name{World_oil}
\alias{World_oil}
\title{Yearly production of oil, globally}
\format{
A data frame with 7 rows and numerical variables \code{pressure} and \code{volume}.

A data frame with 7 rows and numerical variables \code{pressure} and \code{volume}.
}
\source{
https://ourworldindata.org/grapher/oil-production-by-country?country=OWID_WRL

https://ourworldindata.org/grapher/oil-production-by-country?country=OWID_WRL
}
\usage{
World_oil

World_oil
}
\description{
Amount of oil in TWh: terra-watt-hours

Amount of oil in TWh: terra-watt-hours
}
\keyword{datasets}
