% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argM.R
\name{argM}
\alias{argM}
\title{Find local extreme points}
\usage{
argM(tilde, domain)
}
\arguments{
\item{tilde}{specification of a function as in makeFun()}

\item{domain}{a domain to search in.}
}
\description{
Find local extreme points
}
\examples{
argM(x^2 ~ x, domain(x=-1:1))

}
