% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{domain}
\alias{domain}
\title{Specify a domain over which a function is to be graphed}
\usage{
domain(...)
}
\description{
This is a convenience function which creates a list with
named elements. It's used as the second argument to \code{contour_plot()}
and \code{slice_plot()}
}
\examples{
\dontrun{
contour_plot(x / y ~ x + y, domain(x=range(-5,5), y=range(1,4)))
}

}
