% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Effective_oxygen.R
\docType{data}
\name{Effective_oxygen}
\alias{Effective_oxygen}
\title{Effective amount of oxygen available at different altitudes}
\format{
A data frame with 30 rows
\itemize{
\item{\code{altitude}} {In feet above sea level}
\item{\code{oxygen}} {Effective oxygen as a proportion of air (percent)}
}
}
\source{
\url{https://wildsafe.org/resources/ask/altitude-safety/oxygen-levels/}
}
\usage{
data(Effective_oxygen)
}
\description{
Air pressure changes with altitude and therefore the amount of oxygen
available changes as well.
}
\keyword{datasets}
