% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorfield_plot.R
\name{gradient_plot}
\alias{gradient_plot}
\title{Plot a vector field}
\usage{
gradient_plot(
  object = NULL,
  formula,
  domain,
  npts = 20,
  color = "black",
  alpha = 0.5
)
}
\description{
Plot a vector field
}
