% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunSurface.R
\name{graphFunSurface}
\alias{graphFunSurface}
\title{Graph a function of two variables as a surface}
\usage{
graphFunSurface(formula, ...)
}
\arguments{
\item{formula}{of the sort accepted by `makeFun()`}

\item{...}{limits for the plot. See examples.}
}
\description{
Graph a function of two variables as a surface
}
\examples{
graphFunSurface(
  sin(2*pi*t/10)*exp(-.2*x) ~ t & x, 
  tlim=range(0,20), xlim=range(0,10))

}
