% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline.R
\name{make_tilde_inline}
\alias{make_tilde_inline}
\alias{simplify_fun}
\alias{simplify_mult}
\alias{inline_expr}
\alias{replace_arg_in_expr}
\title{Turn a 1-line function into an inline formula}
\usage{
make_tilde_inline(tilde)

simplify_fun(fun)

simplify_mult(ex)

inline_expr(ex, old, new, env)

replace_arg_in_expr(ex, old, new)
}
\arguments{
\item{tilde}{a tilde expression as used by \code{makeFun()}.}

\item{fun}{a function, such as produced by \code{makeFun()}.}

\item{ex}{an expression}

\item{old}{the name to be replaced as produced by as.name()}

\item{new}{the name to be substituted in for arg_name}

\item{env}{the environment for the expression}
}
\description{
Turn a 1-line function into an inline formula

apply simplify_mult() to a function to create a new function.

Try to reduce multiplication in a formula
}
\examples{
infer_RHS(x^2 ~ x) # Formula is unchanged
}
