% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Robot_stations.R
\docType{data}
\name{Robot_stations}
\alias{Robot_stations}
\title{Waypoints on a path in two dimensions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 3 columns.
}
\usage{
Robot_stations
}
\description{
Used as an example in \emph{MOSAIC Calculus} in the "Data-driven functions" chapter.
}
\details{
A data frame with \code{x} and \code{y} components
}
\keyword{datasets}
