% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shapes.R
\docType{data}
\name{Blob1}
\alias{Blob1}
\alias{Blob2}
\alias{Blob3}
\alias{Blob4}
\title{Shapes used in moment of inertia calculations}
\format{
Data frames with coordinates \code{x} and \code{y} as columns

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 200 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 200 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 200 rows and 2 columns.
}
\usage{
Blob1

Blob2

Blob3

Blob4
}
\description{
These are used in a handful of examples in the \emph{MOSAIC Calculus} text.
}
\keyword{datasets}
